implement product details page

+++ app/src/main/java/com/example/bookworm/Product.java
@@ -4,14 +4,14 @@ public class Product {
     private String title;
     private String description;
     private String imgUrl;
-    private int price;
+    private double price;
 
     // Default constructor (no-argument constructor) required for Firebase deserialization
     public Product() {
         // Default constructor required by Firebase Realtime Database
     }
 
-    public Product(String title, String description, String imgUrl, int price) {
+    public Product(String title, String description, String imgUrl, double price) {
         this.title = title;
         this.description = description;
         this.imgUrl = imgUrl;
@@ -42,11 +42,11 @@ public class Product {
         this.imgUrl = imgUrl;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
